Add files via upload
this dataset consists of facts pertaining movie information such as the movie names, directors, ratings runtime, actors and genres. in this project, i firstly analysed the data to get a clear understanding of my data and then i built an unsupervised learning model that recommends movies based on your favourite movie. This model uses various features such as Genre, director, movie name and ratings to reccommend a movie. i also included a streamlit app so that we can use the model.